bot_vision_suite/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bot-vision-suite"
version = "1.0.0"
authors = [
    {name = "Automation Suite Developer", email = "matheuszwilkdev@gmail.com"},
]
description = "Advanced GUI automation with OCR and image recognition"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Testing",
    "Topic :: Desktop Environment",
]
keywords = ["automation", "gui", "ocr", "image-recognition", "rpa", "bot", "desktop-automation"]
dependencies = [
    "PyAutoGUI==0.9.54",
    "pytesseract",
    "opencv-python==4.9.0.80",
    "Pillow>=10.0.0",
    "numpy==1.26.4",
    "pyperclip>=1.8.2"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
]
automation = [
    "keyboard==0.13.5",
    "MouseInfo==0.1.3",
    "setuptools",
    "pandas"
]
ai = [
    "openai>=1.0.0"
]
web = [
    "webdriver-manager>=4.0.0"
]

[project.urls]
Homepage = "https://github.com/automation-suite/bot-vision-suite"
Documentation = "https://bot-vision-suite.readthedocs.io/"
Repository = "https://github.com/automation-suite/bot-vision-suite.git"
"Bug Tracker" = "https://github.com/automation-suite/bot-vision-suite/issues"

[tool.setuptools]
packages = ["bot_vision"]

[tool.setuptools.package-dir]
"" = "."

[tool.black]
line-length = 88
target-version = ['py38']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=bot_vision --cov-report=html --cov-report=term-missing"
